sync: rwlock: Allow API alloc/free functions to change pointer arguments.
The functions rwlock_alloc/free() should take the type 'struct rwlock **'
instead of 'struct rwlock *' so that the pointer can be modified by
memory allocation calls.  This will ensure the rwlock exists on the
heap without the caller needing to allocate memory explicitly (this is
because we expect the caller not to use the stack for thread
synchronization).