makefile: Move DEBUG_FLAGS definition back into top-level makefile.
The DEBUG_FLAGS definition does not need to be included in global.mk
because we will always turn on debugging options when compiling unit
tests, since these were meant to be debugged on failure.  Thus, there
is no need to define DEBUG in unit test files, and debugging options
that are used to compile unit tests can just be specified in the
CFLAGS variable in test/Makefile.