list: linked_list: Fix lost memory in 'linklist_delete()'.
When an item is popped from the list, the 'struct element' wrapper needs
to be freed before returning the data pointer, since the API doesn't use
'struct element' in it's public interface.  Otherwise, the reference will
be lost, and it will never be freed.  Since users won't always want to
capture a deleted reference, move the data-returning functionality
of 'linklist_delete()' to 'linklist_remove()', and have
'linklist_delete()' just delete the element and it's data, freeing both.