list: linked_list: Add typedefs for map and fold function callbacks.
Using typedefs allows the API to force the callback not to modify the
list's internal data, while still allowing the user to create callbacks
that use whatever types are needed; casting the callback function
pointer to any of these typedefs (map_fn_t, foldl_fn_t, and foldr_fn_t)
will automatically convert the function to use 'void *' pointers as
the implementation needs.  This also allows us to deal with lost memory
if the user's callback allocates a new heap variable for the return
value, or to skip that step if the callback used the accumulator
correctly.