list: linked_list: Add 'linklist_drop/take_while()' API functions.
'linklist_drop/take_while()' (corresponding to dropWhile and takeWhile
in Haskell) drop or save elements in a list starting at the beginning
and continuing for as long as the predicate remains true.

In order to test these functions more effectively, change the test
predicate 'pred_gt1()' to 'pred_gte1()' and change it's operation
from (> 1) to (>= 1), which doesn't affect any other test cases.