focs: Add modulo operation that always returns a positive integer.
The mathematically defined range for the modulo operation is
0 ≤ (a mod n) < n.  ISO C99 defines the '%' operator to return the
REMAINDER of a it's quotient, which might be negative depending on
the sign of the types; this behavior is unwanted in situations where
the program must work with modular (looping) array indices.  Add a
macro `mod` that calculates the modulo operation correctly.