list: linked_list: Fix bug in '__delete_after()' for single element.
In '__delete_after()', if there is a single element and 'mark' is NULL,
then the element will not be deleted because the condition
'current == mark' will not be met until the for loop has reached the
end of the list, at which point it is too late to delete the element.
If mark is NULL, the entire list should be dropped anyways, so add
a check to see if mark is NULL before deleting each element.