list: ring_buffer: Fix index to VM address translation function.
The beginning of the buffer is chosen randomly due to ASLR, and this
may cause the modulus that wraps the buffer addresses to be wrong if
we do not subtract the address of the buffer's beginning from the head,
which lets us work with absolute values.
Rename this function to '__rbpos_to_addr()'.