list: ring_buffer: Add `rb_insert()` API function.
The function `rb_insert()` inserts data into a ring buffer at an
existing position.  This means that the ring buffer cannot be empty,
because the index 0 does not exist.  However, if overwrite is
disabled for this data structure, it will open a gap at the given
position to place the data into, effectively changing the size of
the buffer.