hof: Simplify HOF types.
C is not a "pure" programming language, and it's never going to be, so
there's no point in enforcing a pure programming style when it just makes
the types way more complex.  Simplify the HOF function types by having
them modify pointers passed to them instead of returning a new pointer.